*: break out common and versioned code
Split up config/ into some subpackages so each version can be its own
package. Break out common code into it's own pacakge to avoid import
cycles.

Use the new go yaml v3 library instead of my fork since upstream now has
the needed functionality.
