+++ go.mod
@@ -3,9 +3,8 @@ module github.com/ajeddeloh/fcct
 go 1.12
 
 require (
-	github.com/ajeddeloh/go-json v0.0.0-20170920214419-6a2fe990e083
-	github.com/ajeddeloh/yaml v0.0.0-20141224210557-6b16a5714269
+	github.com/ajeddeloh/go-json v0.0.0-20170920214419-6a2fe990e083 // indirect
 	github.com/coreos/go-semver v0.3.0
 	github.com/coreos/ignition/v2 v2.0.0-beta
-	github.com/go-yaml/yaml v2.1.0+incompatible
+	gopkg.in/yaml.v3 v3.0.0-20190502103701-55513cacd4ae
 )
